refactor: moved datasets and ingesting components to pages/data folder
