+++ sites/geohub/src/components/pages/data/ingesting/IngestingDatasetRowDetail.svelte
@@ -4,9 +4,9 @@
 	import { filesize } from 'filesize';
 	import { createEventDispatcher } from 'svelte';
 	import Time from 'svelte-time/src/Time.svelte';
-	import DataPreview from './DataPreview.svelte';
-	import DataPreviewContent from './DataPreviewContent.svelte';
-	import PublishedDatasetDeleteDialog from './PublishedDatasetDeleteDialog.svelte';
+	import DataPreview from '$components/data-upload/DataPreview.svelte';
+	import DataPreviewContent from '$components/data-upload/DataPreviewContent.svelte';
+	import PublishedDatasetDeleteDialog from '$components/pages/data/datasets/PublishedDatasetDeleteDialog.svelte';
 
 	const dispatch = createEventDispatcher();
