+++ sites/geohub/src/routes/(app)/data/[id]/+page.svelte
@@ -2,8 +2,8 @@
 	import { browser } from '$app/environment';
 	import { page } from '$app/stores';
 	import CopyToClipboard from '$components/CopyToClipboard.svelte';
-	import PublishedDataset from '$components/data-upload/PublishedDataset.svelte';
-	import PublishedDatasetOperations from '$components/data-upload/PublishedDatasetOperations.svelte';
+	import PublishedDataset from '$components/pages/data/datasets/PublishedDataset.svelte';
+	import PublishedDatasetOperations from '$components/pages/data/datasets/PublishedDatasetOperations.svelte';
 	import StacExplorer from '$components/stac/StacExplorer.svelte';
 	import { fromLocalStorage, getAccessLevelIcon, storageKeys, toLocalStorage } from '$lib/helper';
 	import type { DatasetFeature, Layer, RasterTileMetadata } from '$lib/types';
